backend/pyproject.toml
[project]
name = "insaights-backend"
version = "0.1.0"
description = "insAIghts backend API"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.9.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt==4.0.1",
  "supabase>=2.0.0",
]

[tool.uvicorn]
app = "app.main:app"
host = "0.0.0.0"
port = 8000
